Added Password Functionality, Improved Directory Structure
