+++ Users/managers.py
@@ -1,4 +1,5 @@
 from django.contrib.auth.models import BaseUserManager
+from .helpers import generate_hash, generate_key
 
 class CustomUserManager(BaseUserManager):
         
@@ -14,7 +15,7 @@ class CustomUserManager(BaseUserManager):
             raise ValueError("Password Field Must Not Be Empty") 
         
         if not decryption_key:
-            raise ValueError("Decrpytion Key Field Must Not Be Empty")
+            raise ValueError("Decryption Key Field Must Not Be Empty")
         
         email = self.normalize_email(email)
         user = self.model(email=email, username=username, decryption_key=decryption_key, **other_fields)
@@ -26,6 +27,8 @@ class CustomUserManager(BaseUserManager):
         
         other_fields.setdefault('is_staff', True)
         other_fields.setdefault('is_superuser', True)
+        
+        decryption_key = generate_hash(generate_key())
         
         if other_fields.get('is_staff') is not True:
             raise ValueError("Staff Status must be set to True")
